Memoize generated flag to prevent the game from grinding to a halt.
I think we will have to do this for all sprites that need to be coloured, for
each player, and hold them in memory. I'm unsure how well this will work out
resource-wise but I believe it's necessary if the game is to run at an
acceptable frame-rate. I'll probably put in a loading screen where all this
initialisation will happen.